requirements.txt
certifi==2019.11.28
chardet==3.0.4
flatten-json==0.1.7
heroku==0.1.4
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
pandas==0.25.3
plaid-python==3.4.0
psycopg2==2.8.4
python-dateutil==1.5
python-dotenv==0.10.3
python-http-client==3.2.1
pytz==2019.3
requests==2.22.0
sendgrid==6.1.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.7
